+++ NEWS.md
@@ -1,8 +1,14 @@
 # SpectriPy 0.2
 
+## Changes in 0.2.1
+
+- Add `filterSpectriPy` method and related parameter objects to perform spectra
+  filtering/processing via matchms in python.
+
 ## Changes in 0.2.0
 
 - Use *matchms* version 0.28.2.
+
 
 # SpectriPy 0.1